--- Login.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<script type="text/javascript" src="scripts.js"></script>
-	<link rel="stylesheet" href="vanilla.css">
-	<title>Proof of Concept</title>
-</head>
-<body>
-
-	<h1 class="title"><a class="navMenu title" href="./index.html">ApprAIse</a></h1>
-
-	<ul>
-		<li><a class="navMenu" href="./About.html">About</a></li>
-		<li><a class="navMenu" href="./Leaderboard.html">Leaderboard</a></li>
-		<li><a class="navMenu" href="./Models.html">Models</a></li>
-		<li><a class="navMenu edge" href="./Login.html">Login</a></li>
-	</ul>
-	<h1>Login</h1>
-
-</body>
-</html>
