+++ frontend/Login.html
@@ -9,7 +9,7 @@
 </head>
 <body>
 
-	<h1 class="title"><a class="navMenu title" href="">ApprAIse</a></h1>
+	<h1 class="title"><a class="navMenu title" href="./index.html">ApprAIse</a></h1>
 
 	<ul>
 		<li><a class="navMenu" href="./About.html">About</a></li>
@@ -17,20 +17,7 @@
 		<li><a class="navMenu" href="./Models.html">Models</a></li>
 		<li><a class="navMenu edge" href="./Login.html">Login</a></li>
 	</ul>
-	<h1>Test API</h1>
-	
-	<button type="button" onclick="get_root()">Submit</button>
-	<div id="test-data"></div>
-
-	<h1>Get Hero</h1>
-	
-	<form id="heroForm">
-		<label for="heroName">Name:</label>
-		<input type="text" name="heroName" id="heroName" placeholder="hero name">
-		<button type="submit">Submit</button>
-	</form>
-	
-	<div id="test-read-hero"></div>
+	<h1>Login</h1>
 
 </body>
 </html>
