--- index.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<script type="text/javascript" src="scripts.js"></script>
-	<link rel="stylesheet" href="vanilla.css">
-	<title>Proof of Concept</title>
-</head>
-<body>
-
-	<h1 class="title"><a class="navMenu title" href="">ApprAIse</a></h1>
-
-	<ul>
-		<li><a class="navMenu" href="./About.html">About</a></li>
-		<li><a class="navMenu" href="./Leaderboard.html">Leaderboard</a></li>
-		<li><a class="navMenu" href="./Models.html">Models</a></li>
-		<li><a class="navMenu edge" href="./Login.html">Login</a></li>
-	</ul>
-	<h1>Test API</h1>
-	
-	<button type="button" onclick="get_root()">Submit</button>
-	<div id="test-data"></div>
-
-	<h1>Get Hero</h1>
-	
-	<form id="heroForm">
-		<label for="heroName">Name:</label>
-		<input type="text" name="heroName" id="heroName" placeholder="hero name">
-		<button type="submit">Submit</button>
-	</form>
-	
-	<div id="test-read-hero"></div>
-
-</body>
-</html>
